requirements.txt
python-dotenv
python-telegram-bot>=20
yt-dlp
pydub
audiomentations
soundfile
numpy<2
pytz
tzlocal<3
google-api-python-client
google-auth-oauthlib
ffmpeg-python
librosa
scipy
tqdm
requests

# AI Tools (External Binaries/CLI recommended for GPU)
# Real-ESRGAN NCNN Vulkan (external binary, place in tools/)
# GFPGAN NCNN Vulkan (external binary, place in tools/)

# ==========================================
# STAGE 2: ADVANCED ENGINE (OPTIONAL)
# Uncomment below ONLY if you have NVIDIA GPU
# and want to use PyTorch models (CodeFormer, RIFE, etc.)
# ==========================================
# torch>=1.13.0
# torchvision>=0.14.0
# basicsr>=1.4.2
# facexlib>=0.2.5
# gfpgan>=1.3.8
# realesrgan>=0.3.0
# tqdm
# lmdb
# pyyaml
# tb-nightly
# future
# addict
# yapf
